separeted logic from routes to controllers
