+++ routes/status.js
@@ -1,8 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const fs = require("fs");
-const users = require("../data/users.json");
-const bcrypt = require("bcrypt");
-require("dotenv").config();
 
+const statusController = require("../controllers/statusController");
+router.route("/").get(statusController.getStatus);
 
+module.exports = router;
